requirements.txt
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Django==2.1.5
django-cors-headers==2.4.0
django-debug-permissions==1.0.0
django-extensions==2.1.5
django-mailgun==0.9.1
django-stubs==1.1.0
djangorestframework==3.9.1
djangorestframework-stubs==0.2.0
drf-generators==0.4.0
idna==2.8
Jinja2==2.10.1
livereload==2.6.1
Markdown==3.1.1
MarkupSafe==1.1.1
mkdocs==1.0.4
mkdocs-material==4.3.1
mypy==0.720
mypy-extensions==0.4.1
pep562==1.0
pip-autoremove==0.9.1
psycopg2==2.8.2
pydotplus==2.0.2
Pygments==2.4.1
pymdown-extensions==6.1
pyparsing==2.4.2
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
six==1.12.0
tornado==6.0.3
typed-ast==1.4.0
typing-extensions==3.7.4
urllib3==1.25.3
uWSGI==2.0.18
webencodings==0.5.1
